# openapi.yaml
swagger: "2.0"
info:
  description: "Botfront API"
  title: "Botfront API"
consumes:
- "application/json"
produces:
- "application/json"
schemes:
- "https"
paths:
...
  "/project/{project_id}/responses":
    get:
      operationId: getAllResponses
      parameters:
      - name: project_id
        in: path
        description: Botfront project ID
        required: true
        type: string
      - name: timestamp
        in: query
        description: timestamp in ms of last update
        required: false
        type: number
      - name: metadata
        in: query
        description: if metadata is set to 1, the metadata field will be included in the response sequence
        required: false
        type: integer
      produces:
        - application/json
      summary: Retrieve all bot responses. If you pass the `timestamp` query param it will compare it with the last update time of the responses. If they have not been modified, a 304 response is returned. This allows to periodically check for updates
      responses:
        default:
          description: Unknown error
        403:
          description: Unauthorized
        200:
          description: OK
          schema:
            $ref: '#/definitions/bot_responses'
        304:
          description: Not modified
...
definitions:
  matching_criteria:
    type: object
    properties:
      nlu:
        type: object
        properties: 
          intent: 
            type: string
            example: greet
          entities:
            type: array
            items:
              type: object
              properties: 
                entity:
                  type: string
                  example: city
                value:
                  type: string
                  example: Paris
...
  sequence:
    description: test
    type: array
    items:
      $ref: '#/definitions/quick_replies'
...
  bot_responses:
    type: object
    properties:
      timestamp:
        type: integer
        example: 1550677361363
      responses:
        type: array
        items:
          type: object
          properties:
            key:
              type: string
              example: utter_something
            value:
              type: array
              items:
                type: object
                properties:
                  sequence:
                    $ref: '#/definitions/sequence'
                  lang:
                    type: string
                    example: 'fr'
            match:
              type: object
              $ref: '#/definitions/matching_criteria'
  quick_replies:
    type: object
    properties:
      text:
        type: string
        example: Something
      buttons:
        type: array
        items:
          type: object
          properties:
            title:
              type: string
              example: "Yes"
            payload:
              type: string
              example: "/basics.yes"
security:
- api_key: []
securityDefinitions:
  # This section configures basic authentication with an API key.
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"